Commit of initial changes to the application. The application does not yet send UDP packets based on the accelerometer data. The main activity has been replaced by one that helps calibrate the limits of the accelerometer.
NOTE: that calibration code does not scale properly. The math needs to be re-wroked.

Also did code for puting the wifi parameters on the sd card. To use, have a microSD card in the phone and hit menu->wifi. It will prompt you to enter the password of whatever WPA network you're using. I haven't added WEP support and probably wont. Also note that this only supports WPA personal passphrase, not WPA key or WPA enterprise as is used at the school. The file gets written to the sdcard at:

/Android/com.ebarch.ipgamepad/data/network.config